requirements.txt
torch
torchvision
numpy
scikit-learn
pyyaml
wandb
tensorboard
medmnist
